ASM-8472 Work around dracut module bug in CentOS 7.2
When trying to build the razor microkernel against CentOS 7.2 the
error message "Dracut module 'dmsquash-live' depends on 'img-lib' in
kernel scriptlet" was thrown when by the kernel post-install
script. This blocked creation of the initrd.img file and led to
a non-functional image being creatd.

The issue is tracked by [RH Bugzilla #1285810][0] and should be fixed
in RH / CentOS 7.3. The workaround is to include tar, gzip, dd and
bash as packages in the microkernel.ks file. That fixes the problem
which was a dependency issue -- the img_lib module didn't declare its
dependencies on those utilities.

[0]: https://bugzilla.redhat.com/show_bug.cgi?id=1285810